Revert "Merge pull request #1119 from mturley/warm-migration-end-to-end"
This reverts commit 7cccd541b075ae910ca19acd03c6438beb204ed6, reversing
changes made to ce193f9ff10c9973a50622a0b85763f9d515dd5f.
