+++ app/javascript/react/screens/App/Overview/OverviewActions.js
@@ -197,32 +197,32 @@ export const toggleScheduleMigrationModal = plan => ({
   payload: plan
 });
 
-export const scheduleCutover = ({ planRequest, cutoverTime }) => dispatch =>
+export const scheduleCutover = payload => dispatch =>
   dispatch({
     type: V2V_SCHEDULE_MIGRATION,
     payload: new Promise((resolve, reject) => {
-      const url = planRequest.href;
+      const {
+        cutoverTime,
+        plan: { id: planId }
+      } = payload;
+      const url = `/api/service_templates/${planId}`;
       const body = {
         action: 'edit',
         resource: {
-          options: {
-            cutover_datetime: cutoverTime
+          config_info: {
+            warm_migration_cutover_datetime: cutoverTime
           }
         }
       };
       return API.post(url, body)
         .then(response => {
           resolve(response);
-          const message =
-            cutoverTime === null
-              ? __('Migration cutover successfully unscheduled')
-              : sprintf(
-                  __('Migration cutover successfully scheduled for %s'),
-                  formatDateTime(response.data.options.cutover_datetime)
-                );
           dispatch({
             type: V2V_NOTIFICATION_ADD,
-            message,
+            message: sprintf(
+              __('Migration cutover successfully scheduled for %s'),
+              formatDateTime(response.data.options.config_info.warm_migration_cutover_datetime)
+            ),
             notificationType: 'success',
             actionEnabled: false
           });
