+++ app/javascript/react/screens/App/Plan/components/PlanRequestDetailList/PlanRequestDetailList.js
@@ -1,10 +1,27 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Button, Grid, FormGroup, Toolbar, DropdownButton, MenuItem } from 'patternfly-react';
-import { REQUEST_TASKS_URL } from '../../PlanConstants';
+import {
+  Button,
+  Icon,
+  Grid,
+  FormGroup,
+  ListView,
+  Popover,
+  OverlayTrigger,
+  Spinner,
+  Toolbar,
+  Tooltip,
+  UtilizationBar,
+  DropdownButton,
+  MenuItem
+} from 'patternfly-react';
+import EllipsisWithTooltip from 'react-ellipsis-with-tooltip';
+import { formatDateTime } from '../../../../../../components/dates/MomentDate';
+import { migrationStatusMessage, REQUEST_TASKS_URL } from '../../PlanConstants';
+import TickingIsoElapsedTime from '../../../../../../components/dates/TickingIsoElapsedTime';
 import ConfirmModal from '../../../common/ConfirmModal';
 import ListViewTable from '../../../common/ListViewTable/ListViewTable';
-import PlanRequestDetailListItem from './PlanRequestDetailListItem';
+import StopPropagationOnClick from '../../../common/StopPropagationOnClick';
 
 class PlanRequestDetailList extends React.Component {
   state = {
@@ -23,18 +40,18 @@ class PlanRequestDetailList extends React.Component {
     }
   }
 
-  downloadLogForTask = (logType, task) => {
+  onSelect = (eventKey, task) => {
     const { downloadLogAction, fetchOrchestrationStackUrl, fetchOrchestrationStackAction } = this.props;
-    if (logType === 'migration') {
+    if (eventKey === 'migration') {
       downloadLogAction(task, 'v2v');
-    } else if (logType === 'wrapper') {
+    } else if (eventKey === 'wrapper') {
       downloadLogAction(task, 'wrapper');
     } else {
-      fetchOrchestrationStackAction(fetchOrchestrationStackUrl, logType, task);
+      fetchOrchestrationStackAction(fetchOrchestrationStackUrl, eventKey, task);
     }
   };
 
-  fetchDetailsForTask = task => {
+  overlayTriggerClick = task => {
     if (!this.props.conversionHosts[task.id] && !task.options.conversion_host_name) {
       this.props.fetchConversionHostAction(this.props.fetchConversionHostUrl, task.id);
     }
@@ -97,6 +114,11 @@ class PlanRequestDetailList extends React.Component {
     }
   };
 
+  taskIsSelected = task => {
+    const { selectedTasksForCancel } = this.props;
+    return selectedTasksForCancel.findIndex(t => t.id === task.id) > -1;
+  };
+
   selectAllInProgressTasks = () => {
     const { incompleteTasks } = this.getCancelSelectionState();
     const { updateSelectedTasksForCancelAction } = this.props;
@@ -129,7 +151,6 @@ class PlanRequestDetailList extends React.Component {
     const { showConfirmCancel } = this.state;
 
     const {
-      plan,
       downloadLogInProgressTaskIds,
       ansiblePlaybookTemplate,
       planRequestTasks,
@@ -142,8 +163,6 @@ class PlanRequestDetailList extends React.Component {
       renderPaginationRow,
       conversionHosts
     } = this.props;
-
-    const isWarmMigration = !!plan.options.config_info.warm_migration;
 
     const { allSelected, noneSelected } = this.getCancelSelectionState();
 
@@ -194,21 +213,227 @@ class PlanRequestDetailList extends React.Component {
         </Grid.Row>
         <div className="main-body-content">
           <ListViewTable className="plan-request-details-list">
-            {filteredSortedPaginatedListItems.items.map(task => (
-              <PlanRequestDetailListItem
-                task={task}
-                isWarmMigration={isWarmMigration}
-                conversionHosts={conversionHosts}
-                downloadLogInProgressTaskIds={downloadLogInProgressTaskIds}
-                ansiblePlaybookTemplate={ansiblePlaybookTemplate}
-                markedForCancellation={markedForCancellation}
-                selectedTasksForCancel={selectedTasksForCancel}
-                handleCheckboxChange={this.handleCheckboxChange}
-                fetchDetailsForTask={this.fetchDetailsForTask}
-                downloadLogForTask={this.downloadLogForTask}
-                key={task.id}
-              />
-            ))}
+            {filteredSortedPaginatedListItems.items.map((task, n) => {
+              let currentDescription = task.options.progress
+                ? migrationStatusMessage(task.options.progress.current_description)
+                : '';
+              if (task.options.prePlaybookRunning || task.options.postPlaybookRunning) {
+                currentDescription = `${currentDescription} (${ansiblePlaybookTemplate.name})`;
+              }
+              let mainStatusMessage = currentDescription;
+              let taskCancelled = false;
+              if (markedForCancellation.find(t => t.id === task.id)) {
+                mainStatusMessage = `${currentDescription}: ${__('Cancel request sent')}`;
+                taskCancelled = true;
+              } else if (task.cancelation_status === 'cancel_requested') {
+                mainStatusMessage = `${currentDescription}: ${__('Cancel requested')}`;
+                taskCancelled = true;
+              } else if (task.cancelation_status === 'canceling') {
+                mainStatusMessage = `${currentDescription}: ${__('Cancelling')}`;
+                taskCancelled = true;
+              } else if (task.cancelation_status === 'canceled') {
+                mainStatusMessage = `${currentDescription}: ${__('Cancelled')}`;
+                taskCancelled = true;
+              }
+              const statusDetailMessage =
+                task.options.progress &&
+                task.options.progress.current_state &&
+                task.options.progress.states &&
+                migrationStatusMessage(task.options.progress.states[task.options.progress.current_state].message);
+
+              let leftContent;
+              if (task.message === 'Pending') {
+                leftContent = (
+                  <ListView.Icon
+                    type="pf"
+                    name="pending"
+                    size="md"
+                    style={{ width: 'inherit', backgroundColor: 'transparent' }}
+                  />
+                );
+              } else if (taskCancelled && task.completed) {
+                mainStatusMessage = `${currentDescription}: ${__('Migration cancelled')}`;
+                leftContent = (
+                  <ListView.Icon
+                    type="fa"
+                    name="ban"
+                    size="md"
+                    style={{ width: 'inherit', backgroundColor: 'transparent' }}
+                  />
+                );
+              } else if (task.completed && !task.completedSuccessfully) {
+                leftContent = (
+                  <ListView.Icon
+                    type="pf"
+                    name="error-circle-o"
+                    size="md"
+                    style={{ width: 'inherit', backgroundColor: 'transparent' }}
+                  />
+                );
+              } else if (task.completed) {
+                leftContent = (
+                  <ListView.Icon
+                    type="pf"
+                    name="ok"
+                    size="md"
+                    style={{ width: 'inherit', backgroundColor: 'transparent' }}
+                  />
+                );
+              } else {
+                leftContent = <Spinner loading />;
+              }
+              const label = sprintf(__('%s of %s Migrated'), task.diskSpaceCompletedGb, task.totalDiskSpaceGb);
+
+              const conversionHostName =
+                task.options.conversion_host_name || (conversionHosts[task.id] && conversionHosts[task.id].name);
+
+              const popoverContent = (
+                <Popover id={`popover${task.id}${n}`} title={mainStatusMessage} className="task-info-popover">
+                  <div>
+                    <div>
+                      <b>{__('Start Time')}: </b>
+                      {formatDateTime(task.startDateTime)}
+                    </div>
+                    <div>
+                      <b>{__('Conversion Host')}: </b>
+                      {conversionHostName}
+                    </div>
+                    {task.log_available && (
+                      <div>
+                        <strong>{__('Log:')}</strong>
+                        <br />
+                        {task.options.virtv2v_wrapper.v2v_log}
+                      </div>
+                    )}
+                  </div>
+                </Popover>
+              );
+
+              return (
+                <ListViewTable.Row
+                  key={task.id}
+                  checkboxInput={
+                    <input
+                      type="checkbox"
+                      disabled={taskCancelled || task.completed}
+                      checked={this.taskIsSelected(task)}
+                      onChange={() => {
+                        this.handleCheckboxChange(task);
+                      }}
+                    />
+                  }
+                  leftContent={leftContent}
+                  heading={task.vmName}
+                  additionalInfo={[
+                    <ListViewTable.InfoItem
+                      key={`${task.id}-message`}
+                      style={{
+                        flexDirection: 'column',
+                        alignItems: 'flex-start',
+                        marginRight: 80,
+                        minWidth: 200
+                      }}
+                    >
+                      <div>
+                        <div style={{ display: 'inline-block', textAlign: 'left' }}>
+                          <span>{mainStatusMessage}</span>
+                          <div style={{ maxWidth: 250 }}>
+                            <EllipsisWithTooltip>{statusDetailMessage}</EllipsisWithTooltip>
+                          </div>
+                        </div>
+                        &nbsp;
+                        {/* Todo: revisit FieldLevelHelp props in patternfly-react to support this */}
+                        <StopPropagationOnClick>
+                          <OverlayTrigger
+                            rootClose
+                            trigger="click"
+                            onEnter={() => this.overlayTriggerClick(task)}
+                            placement="left"
+                            overlay={popoverContent}
+                          >
+                            <Button bsStyle="link">
+                              <Icon type="pf" name="info" />
+                            </Button>
+                          </OverlayTrigger>
+                        </StopPropagationOnClick>
+                      </div>
+                      <div>
+                        <ListView.Icon type="fa" size="lg" name="clock-o" />
+                        <TickingIsoElapsedTime
+                          startTime={task.startDateTime}
+                          endTime={task.completed ? task.lastUpdateDateTime : null}
+                        />
+                      </div>
+                    </ListViewTable.InfoItem>,
+                    <ListViewTable.InfoItem key={`${task.id}-times`} style={{ minWidth: 150, paddingRight: 20 }}>
+                      <UtilizationBar
+                        now={task.percentComplete}
+                        min={0}
+                        max={100}
+                        description={label}
+                        label=" "
+                        usedTooltipFunction={(max, now) => (
+                          <Tooltip id={Date.now()}>
+                            {now} % {__('Migrated')}
+                          </Tooltip>
+                        )}
+                        availableTooltipFunction={(max, now) => (
+                          <Tooltip id={Date.now()}>
+                            {max - now} % {__('Remaining')}
+                          </Tooltip>
+                        )}
+                        descriptionPlacementTop
+                      />
+                    </ListViewTable.InfoItem>
+                  ]}
+                  actions={
+                    <DropdownButton
+                      id={`${task.id}-${task.descriptionPrefix}_download_log_dropdown`}
+                      title={__('Download Log')}
+                      pullRight
+                      onSelect={eventKey => this.onSelect(eventKey, task)}
+                      disabled={
+                        !task.log_available ||
+                        (downloadLogInProgressTaskIds &&
+                          downloadLogInProgressTaskIds.find(element => element === task.id) &&
+                          !task.options.prePlaybookComplete &&
+                          !task.options.postPlaybookComplete)
+                      }
+                    >
+                      {task.options.prePlaybookComplete && (
+                        <MenuItem
+                          eventKey="preMigration"
+                          disabled={downloadLogInProgressTaskIds && downloadLogInProgressTaskIds.indexOf(task.id) > -1}
+                        >
+                          {__('Premigration log')}
+                        </MenuItem>
+                      )}
+                      <MenuItem
+                        eventKey="migration"
+                        disabled={downloadLogInProgressTaskIds && downloadLogInProgressTaskIds.indexOf(task.id) > -1}
+                      >
+                        {__('Migration log')}
+                      </MenuItem>
+                      <MenuItem
+                        eventKey="wrapper"
+                        disabled={downloadLogInProgressTaskIds && downloadLogInProgressTaskIds.indexOf(task.id) > -1}
+                      >
+                        {__('Virt-v2v-wrapper log')}
+                      </MenuItem>
+                      {task.options.postPlaybookComplete && (
+                        <MenuItem
+                          eventKey="postMigration"
+                          disabled={downloadLogInProgressTaskIds && downloadLogInProgressTaskIds.indexOf(task.id) > -1}
+                        >
+                          {__('Postmigration log')}
+                        </MenuItem>
+                      )}
+                    </DropdownButton>
+                  }
+                  stacked
+                />
+              );
+            })}
           </ListViewTable>
           {renderPaginationRow(filteredSortedPaginatedListItems)}
         </div>
