+++ app/javascript/react/screens/App/common/ListViewTable/ListViewTableRow.js
@@ -5,7 +5,7 @@ import { ListView, Icon, noop } from 'patternfly-react';
 import { ListViewTableContext } from './ListViewTable';
 import StopPropagationOnClick from '../StopPropagationOnClick';
 
-const isInfoItemObject = infoItem => !React.isValidElement(infoItem) && infoItem.child && infoItem.tdProps;
+// TODO implement expandable support
 
 const BaseListViewTableRow = ({
   className,
@@ -86,19 +86,14 @@ const BaseListViewTableRow = ({
         </td>
       )}
       {additionalInfo &&
-        additionalInfo.map((infoItem, index) => {
-          const hasTdProps = infoItem && isInfoItemObject(infoItem);
-          const { child, tdProps } = hasTdProps ? infoItem : { child: infoItem, tdProps: {} };
-          return (
-            <td
-              key={`info-item-${index}`}
-              className={classNames('list-view-pf-main-info', { 'empty-info-item': !infoItem })}
-              {...tdProps}
-            >
-              {child}
-            </td>
-          );
-        })}
+        additionalInfo.map((infoItem, index) => (
+          <td
+            key={`info-item-${index}`}
+            className={classNames('list-view-pf-main-info', { 'empty-info-item': !infoItem })}
+          >
+            {infoItem}
+          </td>
+        ))}
       {actions && (
         <td>
           <ListView.Actions>
@@ -142,7 +137,6 @@ const BaseListViewTableRow = ({
 
 BaseListViewTableRow.propTypes = {
   ...ListView.Item.propTypes,
-  additionalInfo: PropTypes.arrayOf(PropTypes.oneOfType([PropTypes.node, PropTypes.object])),
   expanded: PropTypes.bool.isRequired,
   toggleExpanded: PropTypes.func.isRequired
 };
@@ -170,9 +164,6 @@ class ListViewTableRow extends React.Component {
           isSmallViewport ? (
             <ListView.Item
               {...props}
-              additionalInfo={props.additionalInfo.map(
-                infoItem => (isInfoItemObject(infoItem) ? infoItem.child : infoItem)
-              )}
               initExpanded={expanded}
               onExpand={this.toggleExpanded}
               onExpandClose={this.toggleExpanded}
@@ -187,8 +178,7 @@ class ListViewTableRow extends React.Component {
 }
 
 ListViewTableRow.propTypes = {
-  ...ListView.Item.propTypes,
-  additionalInfo: PropTypes.arrayOf(PropTypes.oneOfType([PropTypes.node, PropTypes.object]))
+  ...ListView.Item.propTypes
 };
 
 ListViewTableRow.defaultProps = {
