--- app/javascript/react/screens/App/common/warmMigrationConstants.js
@@ -1,1 +1,0 @@
-export const PRECOPYING_DISKS = 'precopying_disks';
